renamed channel_model.dart to radio_channel_model.dart
