+++ lib/models/radio_channel_model.dart
